% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aslist.R
\name{as.list.R6}
\alias{as.list.R6}
\title{Create a list from an R6 object}
\usage{
\method{as.list}{R6}(x, ...)
}
\arguments{
\item{x}{An R6 object.}

\item{...}{Other arguments, which will be ignored.}
}
\description{
This returns a list of public members from the object. It simply calls
\code{as.list.environment}.
}
