% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/completion.R
\name{.instanceCompletions}
\alias{.instanceCompletions}
\title{Given an R6 generator object, provide completions for R6 objects}
\usage{
.instanceCompletions(generator, group = c("public", "private"),
  operator = "$")
}
\arguments{
\item{generator}{An R6 generator object.}

\item{group}{Which group of completions to supply (public or private members).}

\item{operator}{The operator for which to provide completions. R6 only uses
  \code{$} and \code{[[}.}
}
\description{
This function is meant to be used by interactive R programming environments.
}
\keyword{internal}

