% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is.R6}
\alias{is.R6}
\alias{is.R6Class}
\title{Is an object an R6 Class Generator or Object?}
\usage{
is.R6(x)

is.R6Class(x)
}
\arguments{
\item{x}{An object.}
}
\value{
A logical value.
\itemize{
\item{\code{is.R6Class} returns \code{TRUE} when the input is an R6 class
generator and \code{FALSE} otherwise.}
\item{\code{is.R6} returns \code{TRUE} when the input is an R6 object and
\code{FALSE} otherwise.}
}
}
\description{
Checks for R6 class generators and R6 objects.
}
\examples{
class_generator <- R6Class()
object <- class_generator$new()

is.R6Class(class_generator)
is.R6(class_generator)

is.R6Class(object)
is.R6(object)
}
