% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6-package.R
\docType{package}
\name{R6-package}
\alias{R6-package}
\title{R6: Encapsulated Classes with Reference Semantics}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Creates classes with reference semantics, similar to R's built-in reference classes. Compared to reference classes, R6 classes are simpler and lighter-weight, and they are not built on S4 classes so they do not require the methods package. These classes allow public and private members, and they support inheritance, even when the classes are defined in different packages.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://r6.r-lib.org}
  \item \url{https://github.com/r-lib/R6}
  \item Report bugs at \url{https://github.com/r-lib/R6/issues}
}

}
\author{
\strong{Maintainer}: Winston Chang \email{winston@posit.co}

Other contributors:
\itemize{
  \item Posit Software, PBC [copyright holder, funder]
}

}
\keyword{internal}
