% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is.r6}
\alias{is.r6}
\alias{is.r6gen}
\alias{is.r6obj}
\title{Is an object an R6 Class Generator or Object?}
\usage{
is.r6(x)

is.r6obj(x)

is.r6gen(x)
}
\arguments{
\item{x}{Any variable.}
}
\value{
A logical value.
\code{is.r6} returns \code{TRUE} when the input is an R6 class generator or
an R6 object, and \code{FALSE} otherwise.
\code{is.r6gen} returns \code{TRUE} when the input is an R6 class generator
and \code{FALSE} otherwise.
\code{is.r6obj} returns \code{TRUE} when the input is an R6 object and
\code{FALSE} otherwise.
}
\description{
Checks for R6 class generators and R6 objects.
}
\examples{
class_generator <- R6::R6Class()
object <- class_generator$new()
is.r6(class_generator)
is.r6gen(class_generator)
is.r6obj(class_generator)
is.r6(object)
is.r6gen(object)
is.r6obj(object)
}

